pyproject.toml
[project]
name = "code-merger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.0",
    "python-docx>=1.2.0",
]
